overrides: Ensure the first release tag of ruff isn't being skipped
The previous shell script to generate the version list did skip the
first entry in the version list. The sequence "$@" expands to all the
arguments given to a shell script. The xargs invocation feeds the
argument so the first entry in the list will be the 0th argument, the
second entry in the list will be the 1st and so on. The first argument
to the program  is typically the programs name when a file is invoked.
The usage here is slightly different. We do care about the first
argument and thus can't us "$@" since that only cares
about *arguments* (not the programs name).

With this change xargs now invokes a smaller shell script for each of
the entries in the list. This doesn't move out us into a weird corner
with more shell scripting but instead move a bit of the "heavy
lifting" into xargs.